refactor:  serach/[keyword]ページのlogicとview分離
- serach/[keyword]ページのlogicとviewをファイル分けした
- userとsearchページの関数名を修正
